类型文件重命名以及引用添加type关键字
