+++ src/app/background/modify.ts
@@ -6,10 +6,11 @@ import { dirname, join as pathjoin } from "path";
 import { createExParamPromise, getDate } from "../../utils";
 import { createBuffer, createUri, isFileExits, newUri, readFileUri, uriDelete, writeFileUri } from "../../utils/file";
 import { BackgroundConfiguration } from "../../workspace/background";
-import { Disposable, Uri, version } from "vscode";
+import { version } from "vscode";
+import type { Disposable, Uri } from "vscode";
 import { changeLoadState, getNewBackgroundOpacity, imageStoreUri, setBackgroundImageSuccess } from "./utils";
 import { getVersion } from "../../version";
-import { ContentAndUri, info } from "./type";
+import type { ContentAndUri, info } from "./types";
 import { setStatusBarResolve } from "../../utils/interactive";
 import { WError, $rej } from "../../error";
 import { getNodeModulePath } from "../../utils/system";
