+++ src/app/viewImage/types/index.d.ts
