+++ src/time/openPanel.ts
@@ -1,0 +1,7 @@
+
+/**
+ * 打开设置闹钟的操作面板
+ */
+function openOperationPanel () {
+
+}
