+++ src/utils/compression/types/index.d.ts
