+++ src/utils/webview/panel.ts
@@ -1,4 +1,5 @@
-import { ViewColumn, WebviewOptions, WebviewPanel, WebviewPanelOptions, window } from 'vscode';
+import { ViewColumn, window } from 'vscode';
+import type { WebviewOptions, WebviewPanel, WebviewPanelOptions } from 'vscode';
 import { FileMerge } from ".";
 import { errlog } from '../../error';
 import { messageHandle } from './message';
